Fix #12497.
- AST: Stricter type signature to prevent `isToken` from being applied to
  `SyncTokenId` values; they may be constructed as `Token`s or JSON-compatible
  values with the same fields, and `isToken` uses `instanceof` which leaks that
  distinction.
- Alias analysis: Prevent token from being interpreted as a Tree when its type
  has the same numeric value as the looked-for tree type.
