+++ app/ydoc-shared/src/ast/token.ts
@@ -10,12 +10,12 @@ export import TokenType = RawAst.Token.Type
 export import tokenTypes = RawAst.Token.typeNames
 
 /** Whether the given value is a {@link Token}. */
-export function isToken(maybeToken: unknown): maybeToken is Token {
+export function isToken(maybeToken: Token | Ast): maybeToken is Token {
   return maybeToken instanceof Token
 }
 
 /** Whether the given {@link NodeChild} is a {@link NodeChild}<{@link Token}>. */
-export function isTokenChild(child: NodeChild<unknown>): child is NodeChild<Token> {
+export function isTokenChild(child: NodeChild<Token | Ast>): child is NodeChild<Token> {
   return isToken(child.node)
 }
 
@@ -40,11 +40,6 @@ export class Token implements SyncTokenId {
     readonly tokenType_: TokenType | undefined,
     readonly id: TokenId,
   ) {}
-
-  /** The id of this token. */
-  get externalId(): TokenId {
-    return this.id
-  }
 
   /** Construct a {@link Token} without a {@link TokenId}. */
   static new(code: string, type?: TokenType) {
